backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
anthropic==0.26.0
PyGithub==2.3.0
python-docx==1.1.2
reportlab==4.2.0
httpx==0.27.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bleach==6.1.0
email-validator==2.1.1
